refactor: split backend files, move upload/write logic to backend
